P6 Weekly Update Script Using VBA
An Excel VBA script to generate the weekly Excel reports for departments to update. The VBA script is designed to group data by “Roles” in P6 and then create a new sheet for each Role. After all the sheets are created, each sheet is saved a file in two folders. The first folder called “Upload” contains the most recent file. The second folder “Archives” will contain all the files date and time stamped in the file name whenever the report was run.